+++ draft-js-plugins-editor/src/utils/composeDecorators.ts
@@ -1,8 +1,15 @@
 // This code originally has been copied from Recompose
 // https://github.com/acdlite/recompose/blob/master/src/packages/recompose/compose.js
-export default (...funcs) => {
+
+interface ParamFunction {
+  (params: ParamFunction): ParamFunction;
+}
+
+export default function composeDecorators(
+  ...funcs: Array<ParamFunction>
+): ParamFunction {
   if (funcs.length === 0) {
-    return arg => arg;
+    return (arg) => arg;
   }
 
   if (funcs.length === 1) {
@@ -18,4 +25,4 @@ export default (...funcs) => {
     }
     return result;
   };
-};
+}